Move files to views folder
