+++ views/cond2.py
